Breakout huge test file and code cleanup (#153)
* break out DHCP tests and code clean up

* more breakout of aca_test cases

* fix command line option for aca_test

* for demo only

* adding comment and remove demo change

* ACA supports pulsar consumer (#133)

* Add pulsar dependencies and pulsar consumer

* update aca_main.cpp with entry points to pulsar

* rename MessageConsumer to MessagePulsarConsumer

* fix aca_main.cpp

* fix a wrong symbol

* update gs_tests

* update namespace and class name

* a line of code was deleted by mistake

* add producer and test

* fix remaining bugs

* resolve conflicts in Dockerfile

* fix a bug

* fix a bug

* fix a bug

* missing producer in cmakelists

* fix a bug

* break out DHCP tests and code clean up

* more breakout of aca_test cases

* fix command line option for aca_test

* for demo only

* adding comment and remove demo change

* add description for each test suite and update README.md
