+++ README.md
@@ -46,5 +46,4 @@ This main repository of Alcor Control Agent is organized as follows:
 
 ## Notes
 * ovs_control.h and ovs_control.cpp is based on https://github.com/openvswitch/ovs/blob/master/utilities/ovs-ofctl.c
-* aca_async_grpc_server.cpp is based on https://github.com/grpc/grpc/blob/v1.30.0/examples/cpp/helloworld/greeter_async_server.cc
 * aca_grpc.cpp is based on https://github.com/grpc/grpc/blob/v1.30.0/examples/cpp/route_guide/route_guide_server.cc
