--- test/e2e_tests/test
@@ -1,1 +1,0 @@
-../..//mizar/test
